Education Regulation Impact Analyzer (ERIA)/requirements.txt
requests
beautifulsoup4
bs4
python-dotenv
typing
os
fitz
pytesseract
PIL
re
json
yaml
logging
datetime
time
pypdf
pdfplumber
fpdf
nltk
regex
pytest
scikit-learn
joblib
transformers
torch
torchvision
groq
streamlit
graphviz
reportlab